src/requirements.txt
click==8.0.4
Jinja2==3.1.1
Flask==2.1.0
setuptools==61.2.0
spacy==3.2.3
pyresparser==1.0.6
spacytextblob==4.0.0
Corpora==1.0
escape==1.1
pip-tools==6.5.1
beautifulsoup4==4.9.3
selenium==4.1.3
dnspython==2.2.1
pymongo[tls,srv]==4.0.2
sanic==22.3.0
numpy==1.22.3
tensorflow
keras
Pillow
urllib3
pyimgur